Refactor code to reuse in other tests
