+++ tests/integration/api-gateway/gateway-tests/pkg/client/client.go
@@ -1,0 +1,33 @@
+package client
+
+import (
+	"os"
+
+	"k8s.io/client-go/dynamic"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+)
+
+func loadKubeConfigOrDie() *rest.Config {
+	if _, err := os.Stat(clientcmd.RecommendedHomeFile); os.IsNotExist(err) {
+		cfg, err := rest.InClusterConfig()
+		if err != nil {
+			panic(err)
+		}
+		return cfg
+	}
+
+	cfg, err := clientcmd.BuildConfigFromFlags("", clientcmd.RecommendedHomeFile)
+	if err != nil {
+		panic(err)
+	}
+	return cfg
+}
+
+func GetDynamicClient() dynamic.Interface {
+	client, err := dynamic.NewForConfig(loadKubeConfigOrDie())
+	if err != nil {
+		panic(err)
+	}
+	return client
+}
